requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
Django==4.2.27
django-bootstrap-v5==1.0.11
djangorestframework==3.16.1
greenlet==3.2.4
playwright==1.55.0
pyee==13.0.0
soupsieve==2.8.1
sqlparse==0.5.5
typing_extensions==4.15.0
